+++ history.md
@@ -177,3 +177,4 @@
   - *breaking:* `Lifetime` class moved to interfaces project
   - *breaking:* `Instance` related logic moved to interfaces project
   - *breaking:* `AbstractServiceEntry.Instance` property has been replaced with `Instances`
+  - *introduced:* `Service()` recipe explicit args support